requirements.txt
streamlit
opencv-python
pytesseract
Pillow
numpy
deep-translator
gTTS
